add traits for syntax highlighting and contextualized display
these traits are not ideal but are what i can do in rust right now

the initial attempt to add these traits involved a trait providing a
`colorize` function that returned a struct (or tuple) that impl'd
Display but bundled all useful data together. this would be nice because
use would be like:
```
println!("the next instruction is: {}", instr.colorize(settings));
```
but this runs into the same kind of lifetime issues as mentioned in the
ContextRead commit.

additionally complicating is contextualization, which involves a larger
structure than just a copyable color setting - a similar
builder-and-display style use would be nice but involves a third
lifetime (the content, the colors, the contexts) and is not workable.

references could be Rc pointers and this all might work. why not do
that?

moral opposition to cloning Rc pointers when doing any instruction
rendering. might be worth revisiting later if i'm convinced this is too
annoying - it's already close.

meanwhile, it turns out implementing traits on tuples is currently not
possible anyway - tuples are not #[fundamental].

in the future i need to investigate if i can use macros to tie into
`derive`, so that a ShowContextual impl can #[derive(Colorize,
Display)]. those implementations would really just call
ShowContextual.colorize with fewer and fewer parameters non-None.

the blanket impl
```
impl <T, U> Colorize for T where T: ShowContextual<U>
```
is entirely just a relic of one attempt, and can be ignored. this idea
conflicted with coherence rules because someone else could impl Colorize
for some type covered under this blanket impl.