finally replace `stringy` with something usable
addresses must implement a function that returns a struct that is
applicably formattable. particularly because the default Display impl on
primitives is not necessarily desirable, if you want hex. additionally
this allows meaningful Display for complex (eg, not a single primitive)
addresses, such as segmented memory

also expose AddressDisplay to name outside this crate. what an oversight